+++ bin/astrobase.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { serve } from '../src/http/http.server.js';
+
+serve();